fix handling of sign extension and longs in ramsearch engine. fixes observed misbehaviours in n64 ramsearching. not tested real well, but definitely a needed improvement

+++ trunk/BizHawk.Client.Common/tools/Watch.cs
@@ -230,7 +230,7 @@ namespace BizHawk.Client.Common
 			}
 		}
 
-		public static Watch GenerateWatch(MemoryDomain domain, int address, WatchSize size, DisplayType type, bool bigendian, int prev, int changecount)
+		public static Watch GenerateWatch(MemoryDomain domain, int address, WatchSize size, DisplayType type, bool bigendian, long prev, int changecount)
 		{
 			switch (size)
 			{
